allow to add and lookup duplicated paths per host
Since the pathType implementation, host+path cannot be considered unique
because distinct match types can be used to the same path - mostly exact
with begin or prefix. This update adds the ability to find and return
more than one path instance to the same host+path, and also filter by
match type.

An `order` attribute was also added to preserve the order of the
entries, allowing to add more filter criteria to the request, like
header based match used by the gateway api.